+++ src/test/unit/alignsensor_unittest.cc
@@ -21,11 +21,7 @@
 
 extern "C" {
 #include "common/axis.h"
-
-#include "config/parameter_group.h"
-
 #include "sensors/boardalignment.h"
-
 #include "sensors/sensors.h"
 }
 
@@ -43,9 +39,9 @@ extern "C" {
 
 #define DEG2RAD 0.01745329251
 
-static void rotateVector(int32_t mat[3][3], int32_t vec[3], int32_t *out)
+static void rotateVector(int16_t mat[3][3], int16_t vec[3], int16_t *out)
 {
-    int32_t tmp[3];
+    int16_t tmp[3];
 
     for(int i=0; i<3; i++) {
         tmp[i] = 0;
@@ -73,7 +69,7 @@ static void rotateVector(int32_t mat[3][3], int32_t vec[3], int32_t *out)
 //    mat[2][2] =  cos(angle*DEG2RAD);
 //}
 
-static void initYAxisRotation(int32_t mat[][3], int32_t angle)
+static void initYAxisRotation(int16_t mat[][3], int16_t angle)
 {
     mat[0][0] =  cos(angle*DEG2RAD);
     mat[0][1] =  0;
@@ -86,7 +82,7 @@ static void initYAxisRotation(int32_t mat[][3], int32_t angle)
     mat[2][2] =  cos(angle*DEG2RAD);
 }
 
-static void initZAxisRotation(int32_t mat[][3], int32_t angle)
+static void initZAxisRotation(int16_t mat[][3], int16_t angle)
 {
     mat[0][0] =  cos(angle*DEG2RAD);
     mat[0][1] = -sin(angle*DEG2RAD);
@@ -99,18 +95,18 @@ static void initZAxisRotation(int32_t mat[][3], int32_t angle)
     mat[2][2] =  1;
 }
 
-static void testCW(sensor_align_e rotation, int32_t angle)
+static void testCW(sensor_align_e rotation, int16_t angle)
 {
-    int32_t src[XYZ_AXIS_COUNT];
-    int32_t dest[XYZ_AXIS_COUNT];
-    int32_t test[XYZ_AXIS_COUNT];
+    int16_t src[XYZ_AXIS_COUNT];
+    int16_t dest[XYZ_AXIS_COUNT];
+    int16_t test[XYZ_AXIS_COUNT];
 
     // unit vector along x-axis
     src[X] = 1;
     src[Y] = 0;
     src[Z] = 0;
     
-    int32_t matrix[3][3];
+    int16_t matrix[3][3];
     initZAxisRotation(matrix, angle);
     rotateVector(matrix, src, test);
 
@@ -157,18 +153,18 @@ static void testCW(sensor_align_e rotation, int32_t angle)
  * Since the order of flip and rotation matters, these tests make the
  * assumption that the 'flip' occurs first, followed by clockwise rotation
  */
-static void testCWFlip(sensor_align_e rotation, int32_t angle)
+static void testCWFlip(sensor_align_e rotation, int16_t angle)
 {
-    int32_t src[XYZ_AXIS_COUNT];
-    int32_t dest[XYZ_AXIS_COUNT];
-    int32_t test[XYZ_AXIS_COUNT];
+    int16_t src[XYZ_AXIS_COUNT];
+    int16_t dest[XYZ_AXIS_COUNT];
+    int16_t test[XYZ_AXIS_COUNT];
 
     // unit vector along x-axis
     src[X] = 1;
     src[Y] = 0;
     src[Z] = 0;
     
-    int32_t matrix[3][3];
+    int16_t matrix[3][3];
     initYAxisRotation(matrix, 180);
     rotateVector(matrix, src, test);
     initZAxisRotation(matrix, angle);
